refactor: common 패키지명을 payment로 변경
