+++ src/test/java/roomescape/api/ReservationApiTest.java
@@ -25,8 +25,8 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import roomescape.common.PaymentManager;
-import roomescape.common.dto.PaymentError;
+import roomescape.payment.PaymentManager;
+import roomescape.payment.dto.PaymentError;
 import roomescape.exception.custom.reason.payment.PaymentConfirmException;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
